requirements.txt
streamlit>=1.20.0
pdfplumber
pandas
openpyxl
sentence-transformers
faiss-cpu
numpy
requests
tqdm
python-multipart
pyyaml
pdf2image; platform_system!='Windows'
pytesseract; platform_system!='Windows'
